+++ db/gen/coredb/query.sql.go
@@ -5060,7 +5060,7 @@ insert_media_move_active_record(last_updated) as (
     )
     returning last_updated
 ),
-insert_media_add_record(insert_id, active, is_new) as (
+insert_media_add_record(insert_id, active, replaced_current) as (
     insert into token_medias (id, contract_id, token_id, chain, metadata, media, name, description, processing_job_id, active, created_at, last_updated)
     values ($12, $2, $3, $1, $13, $14, $15, $16, (select id from insert_job), $11,
         -- Using timestamps generated from insert_media_move_active_record ensures that the new record is only inserted after the current media is moved
@@ -5070,8 +5070,8 @@ insert_media_add_record(insert_id, active, is_new) as (
     on conflict (contract_id, token_id, chain) where active and not deleted do update
         set metadata = excluded.metadata,
             media = excluded.media,
-            name = coalesce(nullif(excluded.name, ''), name),
-            description = coalesce(nullif(excluded.description, ''), description),
+            name = coalesce(nullif(excluded.name, ''), token_medias.name),
+            description = coalesce(nullif(excluded.description, ''), token_medias.description),
             processing_job_id = excluded.processing_job_id,
             last_updated = now()
     returning id as insert_id, active, id = $12 replaced_current
